api(lcc/vsc): expose loss/power factor as double, not float
The legacy C++ binding declared LccConverterStation::getLossFactor,
getPowerFactor and VscConverterStation::getLossFactor as `double`.
iidm4cpp narrowed them to `float` to mirror powsybl-iidm-api 7.x's
signatures, which forces every downstream caller to widen back and
mismatches the old API surface.

The underlying value is still float-precision (powsybl quantizes the
XML attribute to binary32 on load), but exposing `double` matches the
old signature so downstream code can be ported without retyping. Drop
the redundant static_cast<float>(getDouble(...)) round-trips.

Tests: keep float-precision comparisons (EXPECT_FLOAT_EQ on the
returned value cast back to float) so we stay honest about the
precision floor inherited from the upstream API.

+++ include/iidm/LccConverterStation.h
@@ -17,8 +17,8 @@ public:
     std::string getId() const;
     std::string getName() const;
 
-    float getLossFactor() const;
-    float getPowerFactor() const;
+    double getLossFactor() const;
+    double getPowerFactor() const;
 
     Terminal getTerminal() const;
     HvdcLine getHvdcLine() const;
